Add files via upload
Visualizing the feature map and training the SOM

plotmap2.m: Plots the feature map diagram using any neighborhood function. It takes three arguments: the weight matrix ('w'), a row vector of neighborhood parameters ('Dim'), and the number of iterations ('iter'). It connects points representing weight row vectors to points representing direct neighbors.

sqr_ex.m: Initializes parameters for a 2D self-organizing map. It sets parameters such as the number of iterations, dimensions of the map, learning rate, etc. Then it iteratively updates the weights of the neurons based on input data.

ngbhr1.m: Calculates the neighbors of the winning unit in a two-dimensional SOM when the radius is 1. It takes the index of the winning unit ('Indx') and the dimensions of the map ('Dim') as input.

neigh2d.m: Calculates the neighbors of the winning unit in a two-dimensional SOM with a specified radius ('Rds'). It also takes the index of the winning unit ('Indx') and the dimensions of the map (Dim) as input.